Create Chinook
PROBLEM STATEMENT 1 
The Chinook record store has just signed a deal with a new record label, and you've been tasked with selecting the first
three albums that will be added to the store, from a list of four. All four albums are by artists that don't have any tracks in
the store right now - we have the artist names, and the genre of music they produce:
The record label specializes in artists from the USA, and they have given Chinook some money to advertise the new
albums in the USA, so we're interested in finding out which genres sell the best in the USA.
You'll need to write a query to find out which genres sell the most tracks in the USA, and then create a
visualization of that data using Power BI.


PROBLEM STATEMENT 2
Each customer for the Chinook store gets assigned to a sales support agent within the company when they first make a
purchase. You have been asked to analyze the purchases of customers belonging to each employee to see if any sales
support agent is performing either better or worse than the others.
You might like to consider whether any extra columns from the employee table explain any variance you see, or whether
the variance might instead be indicative of employee performance.

PROBLEM STATEMENT 3
The Chinook store is set up in a way that allows customer to make purchases in one of the two ways:
● purchase a whole album
● purchase a collection of one or more individual tracks.
The store does not let customers purchase a whole album, and then add individual tracks to that same purchase (unless
they do that by choosing each track manually). When customers purchase albums, they are charged the same price as if
they had purchased each of those tracks separately.
Management are currently considering changing their purchasing strategy to save money. The strategy they are
considering is to purchase only the most popular tracks from each album from record companies, instead of purchasing
every track from an album.
We have been asked to find out what percentage of purchases are individual tracks vs whole albums, so that
management can use this data to understand the effect this decision might have on overall revenue.